requirements.txt
agno==2.4.8
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apsw==3.51.2.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
Authlib==1.6.7
autocommand==2.2.2
Automat==25.4.16
azure-ai-inference==1.0.0b9
azure-core==1.38.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
bottle==0.13.3
Brotli==1.2.0
brotlicffi==1.2.0.0
btrfsutil==6.17.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
construct==2.10.70
Crawl4AI==0.8.0
cryptography==46.0.4
css-parser==1.0.10
cssselect==1.4.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
Cython==3.2.4
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
executing==2.2.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.128.2
fastjsonschema==2.21.2
fastuuid==0.14.0
faust-cchardet==2.1.19
feedparser==6.0.12
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-adk==1.24.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.136.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
google_search_results==2.4.2
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html2text==2025.4.15
html5-parser==0.4.12
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
humanize==4.15.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
incremental==24.7.2
inflate64==1.0.4
inflect==7.5.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jaconv==0.4.1
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
libmsym==0.2.4
lit==21.1.6.dev0
litellm==1.81.8
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mechanize==0.4.10
meson==1.10.1
mistune==3.2.0
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multivolumefile==0.2.3
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest_asyncio==1.5.9
netifaces==0.11.0
netsnmp-python==1.0a1
networkx==3.6.1
nltk==3.9.2
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.17.0
opencv-python==4.13.0.92
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
OpenTimelineIO==0.18.1
packaging==26.0
pandas==3.0.0
pandas-stubs==2.3.3.260113
pandocfilters==1.5.1
parso==0.8.5
patchright==1.58.0
pdftotext==2.2.2
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
powerline-status==2.8.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py7zr==1.1.2
pyaaf2==1.7.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybcj==1.0.7
pycairo==1.29.0
pychm==0.8.6
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
pygdbmi==0.11.0.0
Pygments==2.19.2
PyGObject==3.54.5
pyhibp==4.2.0
PyJWT==2.11.0
pykakasi==2.3.1.dev0
PyKCS11==1.5.18
pykeepass==4.1.1.post1
pyOpenSSL==25.3.0
pyotp==2.9.0
pyparsing==3.3.2
pyppmd==1.2.0
PyQt6==6.10.2
PyQt6-WebEngine==6.10.0
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
python-distutils-extra==2.39
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.22
python-yubico==1.3.3
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.3
pyzmq==27.1.0
pyzstd==0.19.1
QtPy==2.4.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
Send2Trash==1.8.3
sentry-sdk==2.51.0
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smbus==1.1
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8.3
spglib==2.7.1.dev0+g12355c77f.d20260104
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
stack_data==0.6.4.dev0+g53501eb8b.d20251216
starlette==0.50.0
sympy==1.14.0
TBB==0.2
tenacity==9.1.3
texttable==1.7.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.2
tqdm==4.67.2
traitlets==5.14.3
transformers==5.1.0
triton==3.6.0
Twisted==24.7.0
typeguard==4.4.4
typer==0.21.1
typer-slim==0.21.1
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
unrardll==0.1.7
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
validators==0.28.3
watchdog==6.0.0
wcwidth==0.5.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.46.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.2.4
yt-dlp-ejs==0.4.0
ytmusicapi==1.11.5
zeroconf==0.148.0
zipp==3.21.0
zope.interface==8.2
zstandard==0.25.0
zxcvbn-rs-py==0.3.0
